Bug fixes for coupons being overridden
Server was copying files too fast which caused some files to use the same unique id which was based off of a time stamp. Added milliseconds to the unique id to fix that. Also made some adjustments to how the numbers are compared so a single digit wouldn't get mixed with a double digit starting with the number.